# openapi.yaml
openapi: 3.0.0
info:
  title: Bank API
  version: 1.0.0
  description: Bank API with interoperability features
  contact:
    name: API Support
    email: support@eerovallistu.site
servers:
  - url: /
    description: Current server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              message:
                type: string
                example: Authentication required
    ValidationError:
      description: Input validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      summary: Register a new user
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - fullName
                - email
              properties:
                username:
                  type: string
                  example: jsmith
                password:
                  type: string
                  example: securePass123!
                fullName:
                  type: string
                  example: John Smith
                email:
                  type: string
                  format: email
                  example: john.smith@example.com
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: User registered successfully
        '400':
          $ref: '#/components/responses/ValidationError'
  /users/me:
    get:
      summary: Get current user profile
      tags: [Users]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      fullName:
                        type: string
                      email:
                        type: string
                        format: email
                      createdAt:
                        type: string
                        format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /sessions:
    post:
      summary: Login user
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: jsmith
                password:
                  type: string
                  example: securePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      fullName:
                        type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Invalid credentials
  /sessions/{token}:
    delete:
      summary: Logout user (invalidate session token)
      tags: [Users]
      parameters:
        - in: path
          name: token
          schema:
            type: string
          required: true
          description: Session token to invalidate
      responses:
        '200':
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /sessions/logout:
    delete:
      summary: Logout current user
      tags: [Users]
      responses:
        '200':
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Successfully logged out
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /transactions/b2b:
    post:
      summary: Process incoming transfer (Bank-to-Bank)
      tags: [Bank-to-Bank]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jwt
              properties:
                jwt:
                  type: string
                  example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEiLCJ0eXAiOiJKV1QifQ...
      responses:
        '200':
          description: Transaction processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  receiverName:
                    type: string
                    example: John Smith
                  transactionId:
                    type: integer
                    example: 42
        '400':
          description: Invalid JWT or signature
        '404':
          description: Destination account not found
        '500':
          description: Processing error
...